feat: reasponsible user-page layout
