+++ src/Pages/UserPage/UserPage.module.scss
@@ -1,0 +1,27 @@
+.order {
+  margin-bottom: 30px;
+  padding: 15px 30px;
+  border: 1px solid #f3f3f3;
+  border-radius: 40px;
+
+  @media (max-width: 550px) {
+    padding: 15px;
+  }
+}
+
+.order-items {
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+
+  @media (max-width: 1000px) {
+    grid-template-columns: repeat(3, 1fr);
+  }
+
+  @media (max-width: 770px) {
+    grid-template-columns: repeat(2, 1fr);
+  }
+
+  @media (max-width: 550px) {
+    grid-template-columns: repeat(1, 1fr);
+  }
+}
