+++ src/components/OrderedCard/OrderedCard.module.scss
@@ -1,11 +1,30 @@
 .card {
-  margin-right: 30px;
-  margin-bottom: 30px;
+  margin: 0 auto 30px;
   border: 1px solid #f3f3f3;
   padding: 30px;
   width: 200px;
   border-radius: 40px;
   transition: box-shadow 0.2s ease-in-out, transform 0.2s ease-in-out;
+
+  @media (max-width: 550px) {
+    padding: 15px;
+    width: 100%;
+    display: flex;
+    align-items: center;
+
+    h5 {
+      margin-top: 0px;
+    }
+
+    img {
+      width: 80px;
+      height: 72px;
+    }
+
+    .wrapper-description {
+      margin-left: 20px;
+    }
+  }
 
   &:hover {
     box-shadow: 0px 20px 35px rgba(0, 0, 0, 0.06);
